When `unnest()`ing a list column that doesn't have data frames in it, they are converted to `data_frames()` with `as_df()`. There is a call to `set_names()` that is run many times, which ends up being rather expensive. Using `names<-` instead seems to cut a decent chunk of time off.

The `vec_is()` call is also really expensive, but hopefully that will get better with a native implementation in vctrs. Mainly the `is_null()` checks of `vec_is()` take a while, but at the C level they should be faster

```r
library(tidyr)
n <- 10000

df <- tibble(
  g = 1:n,
  y = rep(list(1), n),
  z = rep(list(1), n),
  w = rep(list(1), n)
)

# master tidyr 
bench::mark(unnest(df, c(y, z, w)), iterations = 100)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest(df, c(y, z, w))    830ms    936ms      1.06    3.83MB     14.1

# with `names<-`
bench::mark(unnest(df, c(y, z, w)), iterations = 100)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest(df, c(y, z, w))    568ms    707ms      1.41    3.83MB     13.1
```