@lionel- 2020-02-18
Can you look into making `set_names()` faster in rlang first, and only then change the implementations elsewhere if it isn't sufficient? This way all packages benefit from the speed up. It probably makes sense to rewrite it in C.